+++ test/libsyncengine/CMakeLists.txt
@@ -62,8 +62,9 @@ set(testsyncengine_SRCS
         propagation/executor/testexecutorworker.h propagation/executor/testexecutorworker.cpp
         propagation/executor/testfilerescuer.h propagation/executor/testfilerescuer.cpp
         integration/testintegration.h integration/testintegration.cpp
-        integration/testintegration_conflicts.cpp
         integration/testintegration_basics.cpp
+        integration/testintegration_conflicts.cpp
+        integration/testintegration_nodeid_reuse.cpp
         # SyncPal
         syncpal/testsyncpal.h syncpal/testsyncpal.cpp
         syncpal/testsyncpalworker.h syncpal/testsyncpalworker.cpp
